move the stores from global stores directory to provider store directory
