+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/aws/aws-mc-basic/AwsBasicReviewStep.tsx
@@ -1,21 +1,23 @@
 // React imports
 import React, { useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
+
 // Library imports
 import { CdsButton } from '@cds/react/button';
 import { CdsIcon } from '@cds/react/icon';
+
 // App imports
-import { AWS_FIELDS, CREDENTIAL_TYPE } from './AwsManagementClusterBasic.constants';
-import { AwsStore } from '../../../../state-management/stores/Store.aws';
-import { ConfigDisplay, ConfigDisplayData } from '../../../../shared/components/ConfigReview/ConfigDisplay';
-import { NavRoutes } from '../../../../shared/constants/NavRoutes.constants';
-import { STORE_SECTION_FORM } from '../../../../state-management/reducers/Form.reducer';
-import { StepProps } from '../../../../shared/components/wizard/Wizard';
+import { AwsStore } from '../store/Aws.store.mc';
 import {
     AwsConfigDisplayBasicOneTimeCredentials,
     AwsConfigDisplayBasicProfileCredentials,
     AwsConfigDisplayDefaults,
 } from './AwsBasicReview.config';
+import { AWS_FIELDS, CREDENTIAL_TYPE } from './AwsManagementClusterBasic.constants';
+import { ConfigDisplay, ConfigDisplayData } from '../../../../shared/components/ConfigReview/ConfigDisplay';
+import { NavRoutes } from '../../../../shared/constants/NavRoutes.constants';
+import { StepProps } from '../../../../shared/components/wizard/Wizard';
+import { STORE_SECTION_FORM } from '../../../../state-management/reducers/Form.reducer';
 
 export function AwsBasicReviewStep(props: Partial<StepProps>) {
     const { awsState } = useContext(AwsStore);
