+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/aws/aws-mc-basic/AwsManagementClusterBasic.tsx
@@ -2,9 +2,9 @@
 import React from 'react';
 
 // App imports
-import { AWS_MC_BASIC_TAB_NAMES } from './AwsManagementClusterBasic.constants';
 import { AwsBasicReviewStep } from './AwsBasicReviewStep';
 import AwsClusterSettingsStep from '../AwsClusterSettingsStep';
+import { AWS_MC_BASIC_TAB_NAMES } from './AwsManagementClusterBasic.constants';
 import ManagementCredentials from '../aws-mc-common/management-credential-step/ManagementCredentials';
 import useAwsDeployment from '../../../../shared/services/awsDeployment';
 import Wizard from '../../../../shared/components/wizard/Wizard';
