+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/vsphere/vsphere-mc-basic/VsphereBasicReview.config.tsx
@@ -1,3 +1,4 @@
+// App imports
 import { CommonConfigTransformationFunctions, ConfigGroup } from '../../../../shared/components/ConfigReview/ConfigGrid';
 import { ConfigDisplayData } from '../../../../shared/components/ConfigReview/ConfigDisplay';
 import { VSPHERE_FIELDS } from '../VsphereManagementCluster.constants';
