+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/vsphere/vsphere-mc-common/ThumbprintDisplay.test.tsx
@@ -1,5 +1,8 @@
+// React imports
 import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
+
+// App imports
 import { ThumbprintDisplay } from './ThumbprintDisplay';
 
 const testServerName = 'TEST_SERVER';
