+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/vsphere/vsphere-mc-common/ThumbprintDisplay.tsx
@@ -1,5 +1,8 @@
-import { CdsControlMessage } from '@cds/react/forms';
+// React imports
 import React from 'react';
+
+// Library imports
+import { CdsControlMessage } from '@cds/react/forms';
 
 export interface ThumbprintDisplayInputs {
     serverName: string;
