+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/providers/aws/Aws.reducer.tsx
@@ -1,10 +1,10 @@
 // App imports
 
 // Aws MC reducer
+import { awsResourceReducerDescriptor } from './AwsResources.reducer';
 import { formReducerDescriptor } from '../../../state-management/reducers/Form.reducer';
 import { groupedReducers } from '../../../shared/utilities/Reducer.utils';
 import { uiReducerDescriptor } from '../../../state-management/reducers/Ui.reducer';
-import { awsResourceReducerDescriptor } from './AwsResources.reducer';
 
 export default groupedReducers({
     name: 'Aws MC reducer',
