+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/unmanaged-cluster/unmanaged-cluster-advanced/UnmanagedClusterAdvanced.tsx
@@ -3,10 +3,10 @@ import React from 'react';
 
 // App imports
 import { UMC_ADVANCED_TAB_NAMES } from './UnmanagedClusterAdvanced.constants';
-import Wizard from '../../../shared/components/wizard/Wizard';
 import UnmanagedClusterSettingsAdvanced from '../unmanaged-cluster-basic/unmanaged-cluster-settings-step/UnmanagedClustersSettingsAdvanced';
 // eslint-disable-next-line max-len
 import UnmanagedClusterNetworkSettings from '../unmanaged-cluster-basic/unmanaged-cluster-network-settings-step/UnmanagedClustersNetworkSettings';
+import Wizard from '../../../shared/components/wizard/Wizard';
 
 function UnmanagedClusterAdvanced() {
     return (
